[Fix] getSearchSuggestion 타이밍 문제 해결 (#142)
