+++ ACON-iOS/ACON-iOS/Presentation/Upload/ViewModel/SpotSearchViewModel.swift
@@ -6,12 +6,17 @@
 //
 
 import Foundation
+import CoreLocation
 
 class SpotSearchViewModel: Serviceable {
+    
+    // MARK: - Properties
     
     var longitude: Double = 0
     
     var latitude: Double = 0
+    
+    let isLocationReady: ObservablePattern<Bool> = ObservablePattern(nil)
     
     let onSuccessGetSearchSuggestion: ObservablePattern<Bool> = ObservablePattern(nil)
     
@@ -27,6 +32,16 @@ class SpotSearchViewModel: Serviceable {
     
     var reviewVerificationErrorType: ReviewVerificationErrorType? = nil
     
+    
+    // MARK: - LifeCycle
+    
+    init() {
+        ACLocationManager.shared.addDelegate(self)
+    }
+    
+    deinit {
+       ACLocationManager.shared.removeDelegate(self)
+    }
     
     func getSearchKeyword(keyword: String) {
         let parameter = GetSearchKeywordRequest(keyword: keyword)
@@ -117,9 +132,26 @@ class SpotSearchViewModel: Serviceable {
         }
     }
     
-    func setCoordinates(_ latitude: Double, _ longitude: Double) {
-        self.latitude = latitude
-        self.longitude = longitude
+    func checkLocation() {
+        ACLocationManager.shared.checkUserDeviceLocationServiceAuthorization()
+    }
+    
+}
+
+// MARK: - ACLocationManagerDelegate
+
+extension SpotSearchViewModel: ACLocationManagerDelegate {
+    
+    func locationManager(_ manager: ACLocationManager, didUpdateLocation coordinate: CLLocationCoordinate2D) {
+
+        DispatchQueue.main.async { [weak self] in
+            guard let self = self else { return }
+            
+            print("성공 - 위도: \(coordinate.latitude), 경도: \(coordinate.longitude)")
+            self.latitude = coordinate.latitude
+            self.longitude = coordinate.longitude
+            self.isLocationReady.value = true
+        }
     }
     
 }
